% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\docType{methods}
\name{const_hazard}
\alias{const_hazard}
\title{Constant hazard function}
\usage{
const_hazard(r, lambda)
}
\arguments{
\item{r}{The current R vector length.}

\item{lambda}{The parameter for the hazard function.}
}
\value{
A vector of the hazard function for the length
of the current R vector.
}
\description{
Hazard function for use with gaussian underlying
distribution.
}
\examples{
H<- const_hazard(10, 1/100)
}
