% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gamesdatacounts}
\alias{gamesdatacounts}
\title{This is data to be included in the package}
\source{
\url{http://www.lrec-conf.org/proceedings/lrec2018/pdf/335.pdf}
}
\description{
Data used in the LREC paper on the 2016 eurogames tweets.
Includes a column with the counts of numbers of tweets.
The columns present in the matrix at the three sentiment
scores: "neg", "neu", and "pos", and an additional
column for the total number of tweets: "counts"
}
\keyword{data}
