% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\docType{methods}
\name{gaussian_init}
\alias{gaussian_init}
\title{Initialize vectors for gaussian probability functions}
\usage{
gaussian_init(init_params = list(m = 0, k = 0.01, a = 0.01, b = 1e-04),
  dims)
}
\arguments{
\item{init_params}{The list of parameters to be used for
initialization}

\item{dims}{the dimensionality of the data}
}
\value{
List of vectors to be used in the iteratively
updating algorithm of parameters describing the
underlying gaussian distribution of the data.
}
\description{
Takes in the desired initialization parameters,
initializes the vectors needed for the gaussian probability
function \code{gaussian_update}
}
