% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlineCPD.R
\docType{methods}
\name{initOCPD}
\alias{initOCPD}
\title{Initialize ocpd object}
\usage{
initOCPD(dims, init_params = list(list(m = 0, k = 0.01, a = 0.01, b =
  1e-04)), initProb = c(gaussian_init))
}
\arguments{
\item{dims}{The dimensions calculated from the first input
data points.}

\item{init_params}{The list of params required to initialize
the underlying distribution model.}

\item{initProb}{The chosen type of underlying distribution.}
}
\value{
oCPD object initialized with initialization settings.
}
\description{
This function initializes the ocpd object.
It returns an ocpd object with no data, but matrixes and
vectors set up to begin adding to throughout the
running of the algorithm.
}
\examples{
empty_ocpd<- initOCPD(1) # initialize bject with 1 dimensions
}
