% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\docType{methods}
\name{negbinpdf}
\alias{negbinpdf}
\title{Calculate Negative-binomial on vector of parameters}
\usage{
negbinpdf(x, a, b)
}
\arguments{
\item{x}{the current data point}

\item{a}{matrix of alpha params}

\item{b}{matrix of beta params}
}
\value{
Matrix of negative binomial pdf values corresponding
 to each possible run length, for use in accompanying poisson
probability functions.
}
\description{
Computes the negative-binomial posterior predictive density
from input parameter vectors
corresponding to each possible run length for the current
time point. Outputs a vector of probabilities for use
in the accompanying poisson functions.
}
