% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocp-package.R
\docType{package}
\name{ocp-package}
\alias{ocp-package}
\title{Bayesian Online Changepoint Detection for Multivariate Data}
\description{
Provides an implementation of Bayesian online changepoint detection. Handles multivariate and missing data. Computes the set of changepoints with highest probability in an online way (updating the results with each incoming point). Also provides post-processing functions with alternative ways to extract changepoints.
}
\author{
Pagotto, Andrea
}
