% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\docType{methods}
\name{poisson_init}
\alias{poisson_init}
\title{Initialize vectors for poisson probability functions}
\usage{
poisson_init(init_params = list(a = 1, b = 1), dims)
}
\arguments{
\item{init_params}{The list of parameters to be used for
initialization}

\item{dims}{the dimensionality of the data}
}
\value{
List of vectors to be used in the iteratively
updating algorithm of parameters describing the
underlying gaussian distribution of the data.
}
\description{
Takes in the desired initialization parameters,
initializes the vectors needed for the poisson probability
function \code{poisson_update}
}
