% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{str.ocp}
\alias{str.ocp}
\title{Object Structure}
\usage{
\method{str}{ocp}(object, ...)
}
\arguments{
\item{object}{the object to show}

\item{...}{(optional) additional arguments, ignored.}
}
\description{
Print out information about the ocpd object.
}
\examples{
simdatapts<- c(rnorm(n = 50), rnorm(n=50, 100))
ocpd1<- onlineCPD(simdatapts)
str(ocpd1)
}
