% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\docType{methods}
\name{studentpdf}
\alias{studentpdf}
\title{Calculate Student PDF on vector of parameters}
\usage{
studentpdf(x, mu, var, nu)
}
\arguments{
\item{x}{the current data point}

\item{mu}{vector of means}

\item{var}{var parameter of student pdf, degrees of freedom}

\item{nu}{nu parameter of student pdf (number of points
so far)}
}
\value{
Vector of student pdf values corresponding to each
possible run length, for use in accompanying gaussian
probability functions.
}
\description{
Computes the student pdf from input parameter vectors
corresponding to each possible run length for the current
time point. Outputs a vector of probabilities for use
in the accompanying gaussian functions.
}
