% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{summary.ocp}
\alias{summary.ocp}
\title{Object Summary}
\usage{
\method{summary}{ocp}(object, ...)
}
\arguments{
\item{object}{the object to summarize}

\item{...}{(optional) additional arguments, ignored.}
}
\description{
Print out ocpd object summary.
}
\examples{
simdatapts<- c(rnorm(n = 50), rnorm(n=50, 100))
ocpd1<- onlineCPD(simdatapts)
summary(ocpd1)
}
